Reimplemented micro to macro; fixed a bug where the macro was wrong.
The problem was that by using a system DFA instead of the system NFA the
former would get "simplified" and some edges that do not exist in the
DFA would appear in the NFA. The fact that some new paths appear would
therefore throw off the macro inference alogrithm.

Requires an update of karak.