Rewrote how integration check is performed.
Integration checks and usage-enforce rules are now performed in a file
per subsystem. Meaning, if there are N instances, the model checker will
need to be invoked N times.

The full behavior diagram is only used if the user writes an LTL.

The model checker is only invoked if there are SPECs to check.